## Dec 13, 2018
* TOC
{: toc }

Feeling a bit overwhelmed, rushed, frenetic today for some reason. Deep breaths, Steve, deep breaths.

### One prototype at a time

[History and future of p4](https://github.com/stevekrouse/futureofcoding.org/issues/86)

Now that I'm settled on Turbine (modulus the 15 issues I created, and many more to come), I'm beginning to think about what I hope to build with Turbine: p4.

I was getting a bit overwhelmed with all my ideas for p4. I made a big list and asked JE for help figuring out what's essential and what I can do without for this prototype:

* Immutable editing
	* content hash like Unison and IPFS
	* do we hash definitions or expressions?
	* do we do work to cannonicalize hashes? (such as commutativity of addition)
	* no delete
	* no edit, only create new thing
	* refacorting
* Detachable GUI Literals
	* Everywhere you have a GUI, you need to be able to put an expression: detachable GUI literals!
	* GUI for every system f language primitive
	* Bootstrap visual editor?!
	* Are all apps just expressions visualized?
	* Expression-based operating system instead of file-based or app-based
* Other ideas
	* A la carte language semantics, like haskell langauge extensions
	* Always running, like spreadsheet or google doc, no run or pause, only open or close
	* infinite canvas (can I still use Turbine...?)
	* exploring which concepts we can move from the language compiler to the editor (no free variables, type inference is just local if not just a type “suggestion”)

He helped me realize that most of these ideas represent *entirely different prototypes*! I can't just combine them all like this. Prototypes explore one new idea at a time. So what's the idea for p4?

JE hit the nail on the head: p4 is an IDE for Turbine. This is particularly funny because p1 was an IDE for JQuery and p2 was an IDE for VueJS! Those both used blockly but that wouldn't help with p4 because blocks solve a different problem than what p4 is trying to solve: improve the PX of DCTP, most directly by _showing live streams as the user interacts with the output_.

On my run a few hours ago, I was thinking about what my data model would be for p4. Then it hit me: Turbine is the data model. (Along with hareactive, io, and jabz.) Those stream data structures *are the thing the p4 user is creating*. 

I *could* start p4 by building a projectional editor for Turbine/hareactive/io/jabz and all of JS. But can I get away with less?

One way to get away with less is to build a UI for the turbine/hareactive/io/jabz functions/combinators and then have users type regular text js where we need functional expressions (similar to aprt.us and pane).

The other way to do *even less* is to allow text to be the interaction model for Turbine and simply start with a devtools that shows the streams. If my main conviction is that seeing the streams will help, let's *just* do that. What's great about this idea is that a stream visualizer is a pre-req for any other p4 end-product anyways, so I might as well just publish it as it's own thing, and then move on afterwards.

In summary, p4 is now the Turbine devtools project, in the spirit of the CycleJS devtools.

Wait... because this feels reachable (I can imagine myself achieving it in early 2019 and it existing), it makes me wonder more about next steps...

* Do I try to take up the Turbine mantel and get others to use it?
* Do I go towards the structured editor idea?
* Do I experiment with one of the other ideas listed above?

All valid answers. The thing is, whatever I do, I need a UI library I love, and that library needs a stream visualizer, so build that first. And then we'll talk.

### Donation platform

Before starting the Patreon, I wanted to look around for alternatives. I found Stripe, Paypal and Donationbox allow for subscriptions with much less fees. However, I asked a few people with Patreons and a few of the people who said they'd be my supporters, and both said that I should stick with Patreon, mostly for the branding. It explains itself and gives off the right vibes, which is hard to do with a do-it-your-self platform. Patreon it is!

### Next week

* p4 & Turbine
* Patreon launch
* couple hours of freelance

Maybe something like:

* Monday - Patreon
* Tuesday - Patreon
* Wednesday - p4 & Turbine
* Thursday - p4 & Turbine
* Friday - freelance