## Jan 30, 2019
* TOC
{: toc }

Aaaaaaaand, I'm back! It's been too long since I've done research and a proper log entry. Instead of struggling through getting set up with working on Hareactive/Turbine, I thought it better to wait until I spoke with Simon, and use the waiting time for other work, such as freelancing and the podcast. I've also been obsessed with the Sword of Truth series on audio-book, so that's to blame as well.

### Podcast update 1/30/19

As far as the podcast goes, it's been productive:

* The Tudor Girba episode is ready to go, including the transcript, but I just need to write a summary blurb at the top and then publish and submit to HN.

* The JE episode is edited. It just needs an intro and outtro, transcript and blurb.

* I recorded an conversation with Cyrus Omar

* I am speaking with with Michael Arntzenius *in person* on Saturday!

This will leave me with plenty of podcast editing and publishing to do in February in Rome. (My girlfriend and I are headed to Rome for ~3 weeks for more sunshine in our lives. And Italian food, of course.)

### Balance update 1/30/19

I've been doing a better job of balance than in the past. My only remaining problem is that it's a bit difficult to decide when to do what, and when to switch. I've noticed that I get most of my research work done in the days before meeting with JE, and most of my work for Dark done in the days before meeting with Dark. I don't often meet with FRC, so I sometimes forget about them, but I have two active projects for them that are waiting for whenever I have the time...

I'm glad I did the research on Modern Smalltalks for Dark. It was fustrating at the start, but well worth it at the end. Dreading doing it for mbeddr and TLA+ but I know I'll be thankful afterwards. The plan is to spend all of Friday on Dark stuff.

Yesterday, today, and tomorrow are research so as to have impressive things to show and tell to JE tomorrow.

Next week will be Dark, FRC, and some research, some podcast.

### Call with Simon Friis Vindum

I finally got Simon on the phone! It's been a number of weeks of waiting for him to finish up with his exams. He was wonderful. We chatted for ~2 hours, first about a few issues I opened on his projects, then about the devtool, and finally I shared my screen and he help get me set up in the right code environment and got started adding SVG elements to the Turbine library.

One issue he mentioned with the devtools is that intermediate stream variable names don't get captured in any data structure, so people will have to figure those out from context.  

We also talked a bit about how push-based might has a memory leak problem because parents have references to children, that otherwise would be garbage collected.  

### Turbine devtool

Finally the rubber is hitting the road! It's awesome to see this coming together in a real live project at long last. Simon and I came up with a plan for moving forward:

1. Add SVG elements to Turbine. Done, [pending fixes and merge](https://github.com/funkia/turbine/pull/102).
2. Create a visualization for a simple number stream. [Done this morning.](https://github.com/stevekrouse/turbine/blob/f05807fcbf4abcb3eeb31234eb723c5ee9f90fca/examples/simple/index.ts)
3. Visualize push-based stream (label between intervals)
4. Visualize pull-based stream (constantly sample and maybe graph)
5. Visualize non-numbers
6. Visualize data structure dependencies, starting with map...
7. Visualize higher order...
8. Visualize cyclical...
9. Visualize Turbine/HTML-based streams
10. Use the devtool to finish 7GUIs (add types?), and submit a pull req for that.
 11. Write a paper/presentation/video about the devtool

I realized this morning about all the important little things I need to figure out now, such as dynamically displaying all the stream data. I settled on the fastest solution I could think of: simple arrow based pan and zoom. I wonder if I should go extreme with it and just make the font of large objects really tiny and you can zoom in to see their properties. That'd work as a placeholder.

I feel myself already wishing to "eject" from a devtool, so I can do cool things, such as when you hit space, it pops open a box where you can search for operations like in Facebook origami. But instead you can type, with autocomplete, various flow names and it'll suggest flow combinators based on their types, and supply default arguments, and show the output flow as you scroll through the options. Or maybe the options are the output flow!

### Ink and Switch

Last week I met with the Ink and Switch team. It seems like it's former Heroku people doing research towards a company a la Dark. It was really fun! I just got off a call with pvh from their team. He showed me their [farm](https://github.com/inkandswitch/farm) software. So cool! He was repeating back my whole mission back at me! He calls it "Personal Software". They are using it at their company - right now it's a dashboard of a todolist, personal updates, a chat thing. It's living! Amazing! It makes me think that the day will come soon when personal software is upon us. I can already imagine moving the Future of Coding Slack to a platform like this... Very cool!

Their underlying technology is similar to dat/beaker/scuttlebutt/ifps. pvh had to explain how the internet works to me, Nat, UPD, WebRTC, etc, for me to follow. It's build on top of a JSON CRDT so that it causes the least conflicts. It's a peer to peer data model, and