## Feb 8, 2019
* TOC
{: toc }

Woke up late, and only got a few hours of work in today. My girlfriend and I have been very active - running all four days we've been in Rome so far!

My DCTP paper was accepted into Salon de Refuge, which is a mixed blessing. It's great that I know have to take the time to make the piece decent, but it's also not what I was hoping to spend my time on right now... I'll talk it over with JE and get his advice on decreasing the scope of the piece so I can get something good by the workshop on April 2nd. I'll also have to prepare a talk too!

### Turbine buttons that make buttons visualized!

My favorite cyclic and higher-order FRP example now has a very fun visual to go along with it:

<img height="900px" src="https://user-images.githubusercontent.com/2288939/52494957-9e906a80-2bcf-11e9-95bb-a4505688183c.png">

### Tufte-style polish

Glen was nice enough to read my last entry and make a very kind and helpful comment: 

<blockquote class="twitter-tweet" data-conversation="none" data-cards="hidden" data-lang="en"><p lang="en" dir="ltr">&quot;These diagrams were initially disappointing.&quot; There is some visual design work you can do to make them clearer. First by organizing things spatially along a grid, and then by using colors and line/font weights to create layers a la Tufte: <a href="https://t.co/ShIvO55I3s">pic.twitter.com/ShIvO55I3s</a></p>&mdash; Glen Chiacchieri (@Glench) <a href="https://twitter.com/Glench/status/1093634656682078210?ref_src=twsrc%5Etfw">February 7, 2019</a></blockquote>
<script async src="https://platform.twitter.com/widgets.js" charset="utf-8"></script>

This is a really great point. My diagrams thus far have been created about as fast as possible while trying to get 80% of the details right. They are basically pencil and paper sketches, but done on the computer. 

While part of me thinks it'd be fun to try and get these sketches pixel perfect, I worry that is a longer task, and I'd be better served right now inspecting the Hareactive/Turbine stream data structures to get a better sense of what I'll be able glean from them to then visualize. Then once I have a better idea of *what's possible*, I may come back and try and make these pictures clearer and more accurate.