## March 5, 2019
Despite being sick, I spent a couple hours on a new diagrams on Tuesday, March 5, which felt great. I was hoping to get 30 hours of work on these this week, but four is better than zero. Maybe I'll have time tomorrow to work on it as well. Yesterday was just being sick, and today is catching up on emails and researching Coda.

Ultimately this was the culmination of Tuesday's work, which built on JE's suggestions to have vertical flows to help the text flow horizontally and Josh Horowitz's suggestion to have "wormholes" to keep things from getting too graphy.

![frame 4](https://user-images.githubusercontent.com/2288939/53961686-1d01fe80-40e1-11e9-92bf-6c4cf2ecab13.png)

I also re-did the button click:

![frame 3](https://user-images.githubusercontent.com/2288939/53961674-17a4b400-40e1-11e9-9c65-4e81191c78c1.png)

And messed around with collapsing the space between flows, but found it difficult to lay out the mapping text above:

![frame 8](https://user-images.githubusercontent.com/2288939/53961656-0a87c500-40e1-11e9-9023-887a8475560b.png)
![frame 7](https://user-images.githubusercontent.com/2288939/53961661-0fe50f80-40e1-11e9-829d-c8554121dd8e.png)

In the response to my tweet about these new diagrams, Antranig pointed me to the bidirectional lens work by Pierce, which is funny because I have his book right next to me. I am wondering how related the view-update database problem is to the child-to-parent update problem is in HTML UI design. I'm a bit worried they are the same problem and I've wasted all this time not realizing it. It would explain why nobody has been able to figure out end-user-created UIs: the view-update problem is critical and unsolved. I'll need to look into this more...