## March 22 2019
Got the flu in SF so it's been a rough week! Getting back into it now. Apologies if the following thoughts are even more scatterbrained than usual... I find myself wanting to re-evaluate all my projects:

Dark - I have done ~30 reviews so far. I think I'm running out of interesting ones, but I've thought that before... I just shot Ellen an email with a few other ideas and floating the idea of publishing the past reviews out there.

First Round - I need to follow up with them here about maybe taking on a new project because the current project is mostly done.

As far as freelancing goes, I may have space to take on new projects, but we'll see what happens with Dark and First Round. Given the sale of TCS, I'm not so worried about the extra time. I'll take it for research and let opportunities come to me slowly as they have been.

I'm glad I stuck it out with the DCTP essay for the Salon. Next week I'll work on the presentation and the following week I'll head out there. I think it'll be a worthwhile exercise. But then I'm going to focus on my research. No more distractions! The Salon definitely was a distraction from research for the past two months. But ultimately I learned a lot of criticisms of the DCTP paradigm which is a great thing.

As far as the podcast goes, I'm happy with the status quo. Still have a large backlog of interviews, but I'll get through them at some point. 

The Slack group continues to chug along. I keep teasing moving to another platform, such as Zulip, but nothing ever feels worth it on a closer look.

As far as my research goes... I feel a bit tired of the reactive/stream thread I've been pulling on for years now, but at the same time I do think it's important to get to the end of it. I think there's something important about the "view update problem" synergy that I struck on recently, but still need to look into...

Part of me wants to zoom out and go back into a "reading mode". I almost want to pick another problem, instead of finding the right UI paradigm and solving the GUI GUI creator problem. I also want to finish TaPL and the Bob Harper textbook. And all the links I put into my trelloboard of things to research.

But let's dig into my current status of gametrail, the reactive streams visualization/devtool project:

* I built a bunch of 7GUIs in turbine/hareactive
* I did some initial mockups
* I built a simple first live visualization
* I did another mockup, but with less graph layout and vertical streams

Things that are bumming me out:

* my mockups aren't that clear
* for complex apps, I fear the mockups will be even worse
* hareactive/turbine continues to change their api
* progress is slow
* I worry that the cyclic view update problem is more key than I thought, leading to something like the elm architecture in those cases

As far as research next steps goes:

Talk with JE about the view update / bidirectional transformation problem, because I think that's key. Check out the twitter links, jump into the bidirectional literature. In particular, I am wondering about if there's a way to specify write scopes granularity while still feeling elegant: http://www.cs.pomona.edu/~michael/papers/ugrad_thesis.pdf

As well as working on more mockups. While it feels like I've spent a while on these mockups, I need to get into the design mindset of cranking them out before giving up. I think 100 hours is a good minimum to set here before giving up.

One other thing on my mind: while I do believe in the power of visualizations, I wonder if they're ultimately necessary here. One of the main problems of stream is layout out a list and then doing something with that complex stream, and I can think of other tools to help with that than visualizations, such as an editor that can suggest actions on streams or just a better API for lists of HTML elements.

Also, especially after seeing how quickly the Turbine/hareactive API is changing, I am much less excited about actually building a devtool for it. It seems much more reasonable to a bunch of mockups and maybe a BV-like demo that explores the UI of it and then do a write up or something. I'm not sure what the next step would be after that, but I don't think it's super important to figure that out at the moment. I'm content to spend the next month or two *actually* focused on getting some FRP streams visualized.

Before I started writing today, I felt like quitting this research thread as being too difficult but now I find myself as excited as ever to get back to it. Where I used to feel weighed down by the difficulty, I now see it as an opportunity to do truly original and important work. Where other opportunities are easier and lend themselves to quicker turnaround and payoffs, this work is key. However in the future, I may try to be a bit more strategic about which problem I focus on. But I worry that it's impossible to really guess these things from the outside. Most problems get harder the closer you get to them.