## April 2019 Pivot
* TOC
{: toc }

As I went on and on about in this log yesterday, I think it's time for a zoom out and pivot. The main reason this makes sense is that gametrail, which masqueraded as simple was actually solving too many problems all at once. Another very key reason is that gametrail was just too out-of-nowhere to communicate what I was trying to accomplish here. Even if I did climb the gametrail mountain and make something ok, nobody would understand why that is a big deal.

In my convo with JE yesterday, we agreed that a better "next project" for me to work on now would be to "take aprt.us and add Conal Elliot to it." Part of why JE liked this idea so much is that JE's Subtext directly and explicitly inspired aprt.us -- he even spoke with Toby about. This means that this is likely the first real time JE and I will co-author a paper/talk! 

Because the aprt.us/Conal mashup virtual anyone, another way to put it is add Conal Elliot to spreadsheets. From this perspective, this logic is building directly on the talk I just gave at the Salon, which I hope to put online real soon (tomorrow).

### Toby Research Proposal

Next steps on my aprt.us/Conal pivot... JE suggested that it would be a good idea to reach out to Toby with what I'm trying to accomplish and get his perspective and, if possible, buy-in. In fact, we think that'd it be a great target for my next next step. Basically, my next step is to put together a "research proposal" of what I'm trying to accomplish by adding Conal to aprt.us, my game plan to do it, what technologies I am considered,  and what challenges I expect. As  far a research I need to do to make the gameplan, I want to do some mockups in figma, as well as poking around the aprt.us code for an hour or so.

As far as gameplan goes:

* start with time
* then higher orders
* then cycles
* then submit to LIVE 2019

One thing that's also important to add is the strange ordered-tree-ness of HTML. I don't know where this fits in the gameplan, but it's a real challenge (maybe the biggest challenge), so I should decide if I am including it in this experiment towards the beginning (as it's main risk), or leaving it out of this iteration entirely to simply things.

<blockquote class="twitter-tweet" data-lang="en"><p lang="en" dir="ltr">HTML is masquerading as a boring old tree when it&#39;s in fact a much more complicated species of &quot;ordered tree&quot; <a href="https://t.co/cC3T6ELr6M">https://t.co/cC3T6ELr6M</a></p>&mdash; Steve Krouse (@stevekrouse) <a href="https://twitter.com/stevekrouse/status/1115940158191411201?ref_src=twsrc%5Etfw">April 10, 2019</a></blockquote>

### Next days (April 10, 2019)

* Rest of today is organizing todoist (where I moved after Google Inbox shut down) and then (hopefully) publishing Hillel's podcast. Maybe I'll go on a run in Hyde Park because it's not often this sunny in London.
* Tomorrow is mostly working on podcasts (Cyrus Omar, and maybe finishing up Hillel)
* Friday is spent with my parents and a meeting with Dark and FRC
* I won't work this weekend as my parents are in town

Next week my priorities will be:

* Research pitch (8 hours)
* Podcasts (4 hours)
* Prep for JE (and maybe surprise lunch), and then spending the 19th collaborating with him on my project and his

And then I fly to NY for almost a month, where my priorities will be:

* research
* Dark
* throw a meetup
* see all my NY friends