## April 30, 2019
* TOC
{: toc }

This log entry is way overdue. My head feels like it's going to explode given all the mushy thoughts trying to not be forgotten. I gotta do a better job of logging more often and in smaller increments instead of saving them up for these massive dumps.

### Alan Lunch notes

To recap, my lunch with Alan two weeks ago was ridiculous. I was also surprised to see my casual notes go slightly viral. I got way more people reaching out to me to thank me for the notes than in the past. Maybe it's because the notes had a more personal feel? Or maybe it's because of Alan's reputation?

As I should've expected, Alan saw the notes and wasn't happy with how I published them so roughly, and without giving him a chance to clarify things. In particular, he didn't like how I published the photo of the notes its unedited transcription. I removed it, which was sad because it was many peoples' favorite part. However he was very nice to work with me to get the other parts of the notes more in line with what he meant instead of my (mis)impressions.

### Jonathan Edwards London meetings

I got to spend ~4 hours with Jonathan Edwards on 4/20 to discuss the future of both of our research, but we mostly focused on my stuff. (In short, he's considering pivoting from a more general purpose language to one with a more limited focus, such as data science.)

For my own research direction, we zoomed out even further from the [aprt.us+Conal](https://futureofcoding.org/log#toby-research-proposal) research proposal I was working on. For starters, I put out there the idea to put all of my research threads on hold for the next couple years and instead focus on being more of a "community manager / podcaster" full time. JE did not like this idea at all. (Very few people do. It makes my girlfriend and family bug their eyes out.) JE asked me if I'd be ok just supporting others who get all the glory and don't even give me a head nod. That doesn't sound great, but it really does feel like our community could use a full time (or maybe multiple) people organizing conversation... Maybe over time this person will manifest themselves and I won't have to self-sacrifice to make it happen.

So putting that aside, JE asked me about what my research goals are. I am happy to report that my mission and thesis from the /about page have stayed remarkably stable over the last year or so. Either this means I am burying my head in the sand (which seems unlikely given all the research, conversations, podcasts, etc I do) or it is holding up over time.

> The mission of this project is to  _enable all people [[1](https://futureofcoding.org/about#1)] to modify the software they use in the course of using it_.

JE pointed out that this focus on modification brings up thoughts of simple plugin systems of simplistic customization. Of course, that's the opposite of what I'm going for, so it'd be nice to better convey that distinction. I am committing a draft of this research statement in this commit. Maybe that content will flow into a new version of /about at some point soon... I owe JE a version of this by tomorrow, but Friday is the final deadline for him to review the first draft.

### Last week

Last week I mostly did personal errands, travel,  personal finances, family and NY-friends time, and a podcast interview with Jennifer Jacobs. I got a few hours of research in, but not as much as hoped.

### Juan Benet call 4/27/18

One recent highlight was that on Saturday I spoke to Juan Benet for 90 minutes! If you've been following, you'll know this is a big deal. I see him as a modern day Alan Kay, or a younger Elon in some ways. 

He tentatively agreed to come on the podcast to talk about meta Alan-Kay-style things, such as what makes R&D labs work and so important, and why the internet is great but needs work, and his plans for Protocol Labs, what they're working on now, and any thoughts he has on programming languages and interfaces, such as Dyanmicland, VR/AR, AI, and brain-computer interfaces.

He also mentioned that he'd be interested in collaborating on an in-person, recorded conference on some of these topics. I'd come up with the agenda and people and his team would find the space and fly people out. Sounds too good to be true!

### From Todoist to Workflow

Google Inbox wasn't perfect but the simplicity of a single list for email and tasks, and then being able to push things out of the list with a snooze was great. I guess I can replicate that by just emailing tasks to myself given that Gmail has snooze...

I spent 45 min this morning migrating from Todoist to Workflowy. I'm using #tags for categories, organizing little tasks by when I plan to do them hierarchically, and larger tasks get their own tree but are also placed in a date/time. We'll see how this works...

### Overwhelmed 4/30/19

I've been saying for ~6 months now that I have too many projects going on. It's hard to keep track. Maybe workflowy will help. I think a tool is the right answer but I don't have the bandwidth now to build something...

Aidan Cunniffe and Dev suggested that I leverage others. For example, I could have someone do a guest interview on the podcast. I asked Ivan Reese and he seems to be interested... I could also probably nominate people to help "run" the Slack, but it's a bit strange because it's mostly self-sufficient. Aidan mentioned again that he'd be interested in running a sort of online meetup for people to demo their projects, so maybe I could have a "Future of Coding Community Board" of people interested in planning these sorts of things and people can divy up tasks, etc...?

Of course the easiest way to feel less overwhelmed is to quit or stop doing things. For example, it may not make as much sense for me to take on more projects at FRC given that I may also be taking on coding work for Private Prep (the people who bought The Coding Space) for WoofJS... Maybe eventually I could get more sponsorship to stop fundraising entirely, but that feels like a risky move to me... Maybe one day when I feel more secure in my "research reputation."

### This week

* send JE the first draft of my research proposal (15 hours)
* follow up with Private Prep with WoofJS proposal
* follow up with Juan Benet
* reading group (7guis)
* podcast with Lane from Coda
* prep for NY meetup
* follow up with Alan Kay's recommendation from lunch
* process curious_reader's notes 
* process cyrus omar notes